requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
decorator==4.4.2
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.2
Pillow==9.2.0
proglog==0.1.10
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.12
whitenoise==6.2.0
